- refactored data refresh code out of opendata.js route file to each of the respective repositories
